fr-google-jukebox/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Digital Lab"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = { version = "^0.109.0", extras = ["all"] }
alembic = "^1.8.0"
psycopg2 = "^2.9.3"
Faker = "^13.11.1"
gunicorn = "^20.1.0"
google-cloud-logging = "^3.0.0"
cryptography = "^37.0.3"
SQLAlchemy = "2.0.25"
httpx = "^0.23.0"
psycopg2-binary = "^2.9.6"
starlette-context = "^0.3.6"
asgi-correlation-id = "^4.2.0"
google-cloud-firestore = "^2.11.1"
sqlmodel = "^0.0.14"
pyjwt = "^2.8.0"
pytest = "^8.0.2"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"
pytest-env = "^1.1.3"
asyncpg = "^0.29.0"
pytest-asyncio = "^0.23.5.post1"
greenlet = "^3.0.3"
isort = "^5.13.2"
google-api-python-client = "^2.135.0"
tenacity = "^8.4.2"
google-auth = "^2.30.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
google-auth-plugins = "^1.0.1"
google-generativeai = "^0.7.2"
vertexai = "^1.60.0"
sendgrid = "^6.11.0"
google-cloud-secret-manager = "^2.20.2"
google-cloud-storage = "^2.18.2"

[tool.poetry.dev-dependencies]
black = "^22.3"
flake8 = "^4.0.1"
isort = "^5.13.2"

[tool.black]
line-length = 99                                                              # override black's default line-length
exclude = '(\.git|\.mypy_cache|\.tox|venv|\.venv|_build|buck-out|build|dist)'

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
env = [
    "ENV=test",
    "SQLALCHEMY_DATABASE_URI=postgresql+asyncpg://postgres:postgres@localhost:5435/jukebox_db_test",
    "SQLALCHEMY_SILENCE_UBER_WARNING=1",
]
asyncio_mode = "auto"
